Function replacement hooks and some GLES compat replacements

+++ Core/MIPS/ARM/ArmJit.cpp
@@ -357,6 +357,11 @@ bool Jit::ReplaceJalTo(u32 dest) {
 		return false;
 	}
 
+	if (entry->flags & (REPFLAG_HOOKENTER | REPFLAG_HOOKEXIT)) {
+		// If it's a hook, we can't replace the jal, we have to go inside the func.
+		return false;
+	}
+
 	// Warning - this might be bad if the code at the destination changes...
 	if (entry->flags & REPFLAG_ALLOWINLINE) {
 		// Jackpot! Just do it, no flushing. The code will be entirely inlined.
@@ -397,13 +402,23 @@ void Jit::Comp_ReplacementFunc(MIPSOpcode op)
 	if (entry->jitReplaceFunc) {
 		MIPSReplaceFunc repl = entry->jitReplaceFunc;
 		int cycles = (this->*repl)();
-		FlushAll();
-		// Flushed, so R1 is safe.
-		LDR(R1, CTXREG, MIPS_REG_RA * 4);
-		js.downcountAmount = cycles;
-		WriteExitDestInR(R1);
+
+		if (entry->flags & (REPFLAG_HOOKENTER | REPFLAG_HOOKEXIT)) {
+			// Compile the original instruction at this address.  We ignore cycles for hooks.
+			MIPSCompileOp(Memory::Read_Instruction(js.compilerPC, true));
+		} else {
+			FlushAll();
+			// Flushed, so R1 is safe.
+			LDR(R1, CTXREG, MIPS_REG_RA * 4);
+			js.downcountAmount = cycles;
+			WriteExitDestInR(R1);
+			js.compiling = false;
+		}
 	} else if (entry->replaceFunc) {
 		FlushAll();
+		gpr.SetRegImm(SCRATCHREG1, js.compilerPC);
+		MovToPC(SCRATCHREG1);
+
 		// Standard function call, nothing fancy.
 		// The function returns the number of cycles it took in EAX.
 		if (BLInRange((const void *)(entry->replaceFunc))) {
@@ -412,21 +427,20 @@ void Jit::Comp_ReplacementFunc(MIPSOpcode op)
 			MOVI2R(R0, (u32)entry->replaceFunc);
 			BL(R0);
 		}
-		// Alternatively, we could inline it here, instead of calling out, if it's a function
-		// we can emit.
 
-		LDR(R1, CTXREG, MIPS_REG_RA * 4);
-		WriteDownCountR(R0);
-		js.downcountAmount = 0;  // we just subtracted most of it
-		WriteExitDestInR(R1);
+		if (entry->flags & (REPFLAG_HOOKENTER | REPFLAG_HOOKEXIT)) {
+			// Compile the original instruction at this address.  We ignore cycles for hooks.
+			MIPSCompileOp(Memory::Read_Instruction(js.compilerPC, true));
+		} else {
+			LDR(R1, CTXREG, MIPS_REG_RA * 4);
+			WriteDownCountR(R0);
+			js.downcountAmount = 0;  // we just subtracted most of it
+			WriteExitDestInR(R1);
+			js.compiling = false;
+		}
 	} else {
 		ERROR_LOG(HLE, "Replacement function %s has neither jit nor regular impl", entry->name);
 	}
-
-	js.compiling = false;
-
-	// We could even do this in the jal that is branching to the function
-	// but having the op is necessary for the interpreter anyway.
 }
 
 void Jit::Comp_Generic(MIPSOpcode op)
